+++ wp-content/plugins/gutenberg/build-module/interactivity-router/index.js
@@ -1,138 +1,76 @@
 import * as __WEBPACK_EXTERNAL_MODULE__wordpress_interactivity_8e89b257__ from "@wordpress/interactivity";
 /******/ var __webpack_modules__ = ({
 
-/***/ 195:
-/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
+/***/ 317:
+/***/ ((module) => {
 
-/* harmony export */ __webpack_require__.d(__webpack_exports__, {
-/* harmony export */   Ub: () => (/* binding */ fetchHeadAssets),
-/* harmony export */   ed: () => (/* binding */ headElements),
-/* harmony export */   yp: () => (/* binding */ updateHead)
-/* harmony export */ });
-/**
- * The cache of prefetched stylesheets and scripts.
- */
-const headElements = new Map();
+module.exports = import("@wordpress/a11y");;
 
-/**
- * Helper to update only the necessary tags in the head.
- *
- * @async
- * @param newHead The head elements of the new page.
- */
-const updateHead = async newHead => {
-  // Helper to get the tag id store in the cache.
-  const getTagId = tag => tag.id || tag.outerHTML;
+/***/ })
 
-  // Map incoming head tags by their content.
-  const newHeadMap = new Map();
-  for (const child of newHead) {
-    newHeadMap.set(getTagId(child), child);
-  }
-  const toRemove = [];
+/******/ });
+/************************************************************************/
+/******/ // The module cache
+/******/ var __webpack_module_cache__ = {};
+/******/ 
+/******/ // The require function
+/******/ function __webpack_require__(moduleId) {
+/******/ 	// Check if module is in cache
+/******/ 	var cachedModule = __webpack_module_cache__[moduleId];
+/******/ 	if (cachedModule !== undefined) {
+/******/ 		return cachedModule.exports;
+/******/ 	}
+/******/ 	// Create a new module (and put it into the cache)
+/******/ 	var module = __webpack_module_cache__[moduleId] = {
+/******/ 		// no module.id needed
+/******/ 		// no module.loaded needed
+/******/ 		exports: {}
+/******/ 	};
+/******/ 
+/******/ 	// Execute the module function
+/******/ 	__webpack_modules__[moduleId](module, module.exports, __webpack_require__);
+/******/ 
+/******/ 	// Return the exports of the module
+/******/ 	return module.exports;
+/******/ }
+/******/ 
+/************************************************************************/
+/******/ /* webpack/runtime/define property getters */
+/******/ (() => {
+/******/ 	// define getter functions for harmony exports
+/******/ 	__webpack_require__.d = (exports, definition) => {
+/******/ 		for(var key in definition) {
+/******/ 			if(__webpack_require__.o(definition, key) && !__webpack_require__.o(exports, key)) {
+/******/ 				Object.defineProperty(exports, key, { enumerable: true, get: definition[key] });
+/******/ 			}
+/******/ 		}
+/******/ 	};
+/******/ })();
+/******/ 
+/******/ /* webpack/runtime/hasOwnProperty shorthand */
+/******/ (() => {
+/******/ 	__webpack_require__.o = (obj, prop) => (Object.prototype.hasOwnProperty.call(obj, prop))
+/******/ })();
+/******/ 
+/************************************************************************/
+var __webpack_exports__ = {};
 
-  // Detect nodes that should be added or removed.
-  for (const child of document.head.children) {
-    const id = getTagId(child);
-    // Always remove styles and links as they might change.
-    if (child.nodeName === 'LINK' || child.nodeName === 'STYLE') {
-      toRemove.push(child);
-    } else if (newHeadMap.has(id)) {
-      newHeadMap.delete(id);
-    } else if (child.nodeName !== 'SCRIPT' && child.nodeName !== 'META') {
-      toRemove.push(child);
-    }
-  }
-  await Promise.all([...headElements.entries()].filter(([, {
-    tag
-  }]) => tag.nodeName === 'SCRIPT').map(async ([url]) => {
-    await import(/* webpackIgnore: true */url);
-  }));
+// EXPORTS
+__webpack_require__.d(__webpack_exports__, {
+  o: () => (/* binding */ actions),
+  w: () => (/* binding */ state)
+});
 
-  // Prepare new assets.
-  const toAppend = [...newHeadMap.values()];
-
-  // Apply the changes.
-  toRemove.forEach(n => n.remove());
-  document.head.append(...toAppend);
-};
-
-/**
- * Fetches and processes head assets (stylesheets and scripts) from a specified document.
- *
- * @async
- * @param doc The document from which to fetch head assets. It should support standard DOM querying methods.
- *
- * @return Returns an array of HTML elements representing the head assets.
- */
-const fetchHeadAssets = async doc => {
-  const headTags = [];
-
-  // We only want to fetch module scripts because regular scripts (without
-  // `async` or `defer` attributes) can depend on the execution of other scripts.
-  // Scripts found in the head are blocking and must be executed in order.
-  const scripts = doc.querySelectorAll('script[type="module"][src]');
-  scripts.forEach(script => {
-    const src = script.getAttribute('src');
-    if (!headElements.has(src)) {
-      // add the <link> elements to prefetch the module scripts
-      const link = doc.createElement('link');
-      link.rel = 'modulepreload';
-      link.href = src;
-      document.head.append(link);
-      headElements.set(src, {
-        tag: script
-      });
-    }
-  });
-  const stylesheets = doc.querySelectorAll('link[rel=stylesheet]');
-  await Promise.all(Array.from(stylesheets).map(async tag => {
-    const href = tag.getAttribute('href');
-    if (!href) {
-      return;
-    }
-    if (!headElements.has(href)) {
-      try {
-        const response = await fetch(href);
-        const text = await response.text();
-        headElements.set(href, {
-          tag,
-          text
-        });
-      } catch (e) {
-        // eslint-disable-next-line no-console
-        console.error(e);
-      }
-    }
-    const headElement = headElements.get(href);
-    const styleElement = doc.createElement('style');
-    styleElement.textContent = headElement.text;
-    headTags.push(styleElement);
-  }));
-  return [doc.querySelector('title'), ...doc.querySelectorAll('style'), ...headTags];
-};
-
-
-/***/ }),
-
-/***/ 873:
-/***/ ((module, __webpack_exports__, __webpack_require__) => {
-
-__webpack_require__.a(module, async (__webpack_handle_async_dependencies__, __webpack_async_result__) => { try {
-/* harmony export */ __webpack_require__.d(__webpack_exports__, {
-/* harmony export */   o: () => (/* binding */ actions),
-/* harmony export */   w: () => (/* binding */ state)
-/* harmony export */ });
-/* harmony import */ var _wordpress_interactivity__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(833);
-/* harmony import */ var _head__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(195);
-var _getConfig$navigation;
+;// external "@wordpress/interactivity"
+var x = (y) => {
+	var x = {}; __webpack_require__.d(x, y); return x
+} 
+var y = (x) => (() => (x))
+const interactivity_namespaceObject = x({ ["getConfig"]: () => (__WEBPACK_EXTERNAL_MODULE__wordpress_interactivity_8e89b257__.getConfig), ["privateApis"]: () => (__WEBPACK_EXTERNAL_MODULE__wordpress_interactivity_8e89b257__.privateApis), ["store"]: () => (__WEBPACK_EXTERNAL_MODULE__wordpress_interactivity_8e89b257__.store) });
+;// ./packages/interactivity-router/build-module/index.js
+/* wp:polyfill */
 /**
  * WordPress dependencies
- */
-
-
-/**
- * Internal dependencies
  */
 
 const {
@@ -144,9 +82,10 @@ const {
   parseServerData,
   populateServerData,
   batch
-} = (0,_wordpress_interactivity__WEBPACK_IMPORTED_MODULE_0__.privateApis)('I acknowledge that using private APIs means my theme or plugin will inevitably break in the next version of WordPress.');
-// Check if the navigation mode is full page or region based.
-const navigationMode = (_getConfig$navigation = (0,_wordpress_interactivity__WEBPACK_IMPORTED_MODULE_0__.getConfig)('core/router').navigationMode) !== null && _getConfig$navigation !== void 0 ? _getConfig$navigation : 'regionBased';
+} = (0,interactivity_namespaceObject.privateApis)('I acknowledge that using private APIs means my theme or plugin will inevitably break in the next version of WordPress.');
+// Check if the navigation mode is full page or region based. The only supported
+// mode for now is 'regionBased'.
+const navigationMode = 'regionBased';
 
 // The cache of visited and prefetched pages, stylesheets and scripts.
 const pages = new Map();
@@ -185,13 +124,6 @@ const regionsToVdom = async (dom, {
   const regions = {
     body: undefined
   };
-  let head;
-  if (true) {
-    if (navigationMode === 'fullPage') {
-      head = await (0,_head__WEBPACK_IMPORTED_MODULE_1__/* .fetchHeadAssets */ .Ub)(dom);
-      regions.body = vdom ? vdom.get(document.body) : toVdom(dom.body);
-    }
-  }
   if (navigationMode === 'regionBased') {
     const attrName = `data-${directivePrefix}-router-region`;
     dom.querySelectorAll(`[${attrName}]`).forEach(region => {
@@ -203,7 +135,6 @@ const regionsToVdom = async (dom, {
   const initialData = parseServerData(dom);
   return {
     regions,
-    head,
     title,
     initialData
   };
@@ -211,17 +142,6 @@ const regionsToVdom = async (dom, {
 
 // Render all interactive regions contained in the given page.
 const renderRegions = async page => {
-  if (true) {
-    if (navigationMode === 'fullPage') {
-      // Once this code is tested and more mature, the head should be updated for region based navigation as well.
-      await (0,_head__WEBPACK_IMPORTED_MODULE_1__/* .updateHead */ .yp)(page.head);
-      const fragment = getRegionRootFragment(document.body);
-      batch(() => {
-        populateServerData(page.initialData);
-        render(page.regions.body, fragment);
-      });
-    }
-  }
   if (navigationMode === 'regionBased') {
     const attrName = `data-${directivePrefix}-router-region`;
     batch(() => {
@@ -268,36 +188,9 @@ window.addEventListener('popstate', async () => {
 });
 
 // Initialize the router and cache the initial page using the initial vDOM.
-// Once this code is tested and more mature, the head should be updated for
-// region based navigation as well.
-if (true) {
-  if (navigationMode === 'fullPage') {
-    // Cache the scripts. Has to be called before fetching the assets.
-    [].map.call(document.querySelectorAll('script[type="module"][src]'), script => {
-      _head__WEBPACK_IMPORTED_MODULE_1__/* .headElements */ .ed.set(script.getAttribute('src'), {
-        tag: script
-      });
-    });
-    await (0,_head__WEBPACK_IMPORTED_MODULE_1__/* .fetchHeadAssets */ .Ub)(document);
-  }
-}
 pages.set(getPagePath(window.location.href), Promise.resolve(regionsToVdom(document, {
   vdom: initialVdom
 })));
-
-// Check if the link is valid for client-side navigation.
-const isValidLink = ref => ref && ref instanceof window.HTMLAnchorElement && ref.href && (!ref.target || ref.target === '_self') && ref.origin === window.location.origin && !ref.pathname.startsWith('/wp-admin') && !ref.pathname.startsWith('/wp-login.php') && !ref.getAttribute('href').startsWith('#') && !new URL(ref.href).searchParams.has('_wpnonce');
-
-// Check if the event is valid for client-side navigation.
-const isValidEvent = event => event && event.button === 0 &&
-// Left clicks only.
-!event.metaKey &&
-// Open in new tab (Mac).
-!event.ctrlKey &&
-// Open in new tab (Windows).
-!event.altKey &&
-// Download.
-!event.shiftKey && !event.defaultPrevented;
 
 // Variable to store the current navigation.
 let navigatingTo = '';
@@ -309,7 +202,7 @@ const navigationTexts = {
 const {
   state,
   actions
-} = (0,_wordpress_interactivity__WEBPACK_IMPORTED_MODULE_0__.store)('core/router', {
+} = (0,interactivity_namespaceObject.store)('core/router', {
   state: {
     url: window.location.href,
     navigation: {
@@ -339,7 +232,7 @@ const {
     *navigate(href, options = {}) {
       const {
         clientNavigationDisabled
-      } = (0,_wordpress_interactivity__WEBPACK_IMPORTED_MODULE_0__.getConfig)();
+      } = (0,interactivity_namespaceObject.getConfig)();
       if (clientNavigationDisabled) {
         yield forcePageReload(href);
       }
@@ -425,7 +318,7 @@ const {
     prefetch(url, options = {}) {
       const {
         clientNavigationDisabled
-      } = (0,_wordpress_interactivity__WEBPACK_IMPORTED_MODULE_0__.getConfig)();
+      } = (0,interactivity_namespaceObject.getConfig)();
       if (clientNavigationDisabled) {
         return;
       }
@@ -485,173 +378,6 @@ function a11ySpeak(messageKey) {
   () => {});
 }
 
-// Add click and prefetch to all links.
-if (true) {
-  if (navigationMode === 'fullPage') {
-    // Navigate on click.
-    document.addEventListener('click', function (event) {
-      const ref = event.target.closest('a');
-      if (isValidLink(ref) && isValidEvent(event)) {
-        event.preventDefault();
-        actions.navigate(ref.href);
-      }
-    }, true);
-    // Prefetch on hover.
-    document.addEventListener('mouseenter', function (event) {
-      if (event.target?.nodeName === 'A') {
-        const ref = event.target.closest('a');
-        if (isValidLink(ref) && isValidEvent(event)) {
-          actions.prefetch(ref.href);
-        }
-      }
-    }, true);
-  }
-}
-
-__webpack_async_result__();
-} catch(e) { __webpack_async_result__(e); } }, 1);
-
-/***/ }),
-
-/***/ 317:
-/***/ ((module) => {
-
-module.exports = import("@wordpress/a11y");;
-
-/***/ }),
-
-/***/ 833:
-/***/ ((module, __unused_webpack_exports, __webpack_require__) => {
-
-var x = (y) => {
-	var x = {}; __webpack_require__.d(x, y); return x
-} 
-var y = (x) => (() => (x))
-module.exports = x({ ["getConfig"]: () => (__WEBPACK_EXTERNAL_MODULE__wordpress_interactivity_8e89b257__.getConfig), ["privateApis"]: () => (__WEBPACK_EXTERNAL_MODULE__wordpress_interactivity_8e89b257__.privateApis), ["store"]: () => (__WEBPACK_EXTERNAL_MODULE__wordpress_interactivity_8e89b257__.store) });
-
-/***/ })
-
-/******/ });
-/************************************************************************/
-/******/ // The module cache
-/******/ var __webpack_module_cache__ = {};
-/******/ 
-/******/ // The require function
-/******/ function __webpack_require__(moduleId) {
-/******/ 	// Check if module is in cache
-/******/ 	var cachedModule = __webpack_module_cache__[moduleId];
-/******/ 	if (cachedModule !== undefined) {
-/******/ 		return cachedModule.exports;
-/******/ 	}
-/******/ 	// Create a new module (and put it into the cache)
-/******/ 	var module = __webpack_module_cache__[moduleId] = {
-/******/ 		// no module.id needed
-/******/ 		// no module.loaded needed
-/******/ 		exports: {}
-/******/ 	};
-/******/ 
-/******/ 	// Execute the module function
-/******/ 	__webpack_modules__[moduleId](module, module.exports, __webpack_require__);
-/******/ 
-/******/ 	// Return the exports of the module
-/******/ 	return module.exports;
-/******/ }
-/******/ 
-/************************************************************************/
-/******/ /* webpack/runtime/async module */
-/******/ (() => {
-/******/ 	var webpackQueues = typeof Symbol === "function" ? Symbol("webpack queues") : "__webpack_queues__";
-/******/ 	var webpackExports = typeof Symbol === "function" ? Symbol("webpack exports") : "__webpack_exports__";
-/******/ 	var webpackError = typeof Symbol === "function" ? Symbol("webpack error") : "__webpack_error__";
-/******/ 	var resolveQueue = (queue) => {
-/******/ 		if(queue && queue.d < 1) {
-/******/ 			queue.d = 1;
-/******/ 			queue.forEach((fn) => (fn.r--));
-/******/ 			queue.forEach((fn) => (fn.r-- ? fn.r++ : fn()));
-/******/ 		}
-/******/ 	}
-/******/ 	var wrapDeps = (deps) => (deps.map((dep) => {
-/******/ 		if(dep !== null && typeof dep === "object") {
-/******/ 			if(dep[webpackQueues]) return dep;
-/******/ 			if(dep.then) {
-/******/ 				var queue = [];
-/******/ 				queue.d = 0;
-/******/ 				dep.then((r) => {
-/******/ 					obj[webpackExports] = r;
-/******/ 					resolveQueue(queue);
-/******/ 				}, (e) => {
-/******/ 					obj[webpackError] = e;
-/******/ 					resolveQueue(queue);
-/******/ 				});
-/******/ 				var obj = {};
-/******/ 				obj[webpackQueues] = (fn) => (fn(queue));
-/******/ 				return obj;
-/******/ 			}
-/******/ 		}
-/******/ 		var ret = {};
-/******/ 		ret[webpackQueues] = x => {};
-/******/ 		ret[webpackExports] = dep;
-/******/ 		return ret;
-/******/ 	}));
-/******/ 	__webpack_require__.a = (module, body, hasAwait) => {
-/******/ 		var queue;
-/******/ 		hasAwait && ((queue = []).d = -1);
-/******/ 		var depQueues = new Set();
-/******/ 		var exports = module.exports;
-/******/ 		var currentDeps;
-/******/ 		var outerResolve;
-/******/ 		var reject;
-/******/ 		var promise = new Promise((resolve, rej) => {
-/******/ 			reject = rej;
-/******/ 			outerResolve = resolve;
-/******/ 		});
-/******/ 		promise[webpackExports] = exports;
-/******/ 		promise[webpackQueues] = (fn) => (queue && fn(queue), depQueues.forEach(fn), promise["catch"](x => {}));
-/******/ 		module.exports = promise;
-/******/ 		body((deps) => {
-/******/ 			currentDeps = wrapDeps(deps);
-/******/ 			var fn;
-/******/ 			var getResult = () => (currentDeps.map((d) => {
-/******/ 				if(d[webpackError]) throw d[webpackError];
-/******/ 				return d[webpackExports];
-/******/ 			}))
-/******/ 			var promise = new Promise((resolve) => {
-/******/ 				fn = () => (resolve(getResult));
-/******/ 				fn.r = 0;
-/******/ 				var fnQueue = (q) => (q !== queue && !depQueues.has(q) && (depQueues.add(q), q && !q.d && (fn.r++, q.push(fn))));
-/******/ 				currentDeps.map((dep) => (dep[webpackQueues](fnQueue)));
-/******/ 			});
-/******/ 			return fn.r ? promise : getResult();
-/******/ 		}, (err) => ((err ? reject(promise[webpackError] = err) : outerResolve(exports)), resolveQueue(queue)));
-/******/ 		queue && queue.d < 0 && (queue.d = 0);
-/******/ 	};
-/******/ })();
-/******/ 
-/******/ /* webpack/runtime/define property getters */
-/******/ (() => {
-/******/ 	// define getter functions for harmony exports
-/******/ 	__webpack_require__.d = (exports, definition) => {
-/******/ 		for(var key in definition) {
-/******/ 			if(__webpack_require__.o(definition, key) && !__webpack_require__.o(exports, key)) {
-/******/ 				Object.defineProperty(exports, key, { enumerable: true, get: definition[key] });
-/******/ 			}
-/******/ 		}
-/******/ 	};
-/******/ })();
-/******/ 
-/******/ /* webpack/runtime/hasOwnProperty shorthand */
-/******/ (() => {
-/******/ 	__webpack_require__.o = (obj, prop) => (Object.prototype.hasOwnProperty.call(obj, prop))
-/******/ })();
-/******/ 
-/************************************************************************/
-/******/ 
-/******/ // startup
-/******/ // Load entry module and return exports
-/******/ // This entry module used 'module' so it can't be inlined
-/******/ var __webpack_exports__ = __webpack_require__(873);
-/******/ __webpack_exports__ = await __webpack_exports__;
-/******/ var __webpack_exports__actions = __webpack_exports__.o;
-/******/ var __webpack_exports__state = __webpack_exports__.w;
-/******/ export { __webpack_exports__actions as actions, __webpack_exports__state as state };
-/******/ 
+var __webpack_exports__actions = __webpack_exports__.o;
+var __webpack_exports__state = __webpack_exports__.w;
+export { __webpack_exports__actions as actions, __webpack_exports__state as state };