+++ wp-content/plugins/gutenberg/gutenberg.php
@@ -5,7 +5,7 @@
  * Description: Printing since 1440. This is the development plugin for the block editor, site editor, and other future WordPress core functionality.
  * Requires at least: 6.7
  * Requires PHP: 7.2
- * Version: 20.9.0
+ * Version: 21.0.0
  * Author: Gutenberg Team
  * Text Domain: gutenberg
  *
@@ -13,8 +13,8 @@
  */
 
 ### BEGIN AUTO-GENERATED DEFINES
-define( 'GUTENBERG_VERSION', '20.9.0' );
-define( 'GUTENBERG_GIT_COMMIT', 'b5e34608ea83710a3ebdddd9256398c6b4608746' );
+define( 'GUTENBERG_VERSION', '21.0.0' );
+define( 'GUTENBERG_GIT_COMMIT', '1d56368a95d1042f8a3de91943c477641d057c0f' );
 ### END AUTO-GENERATED DEFINES
 defined( 'GUTENBERG_MINIMUM_WP_VERSION' ) or define( 'GUTENBERG_MINIMUM_WP_VERSION', '6.7' );