+++ wp-content/plugins/gutenberg/lib/experiments-page.php
@@ -176,18 +176,6 @@ function gutenberg_initialize_experiments_settings() {
 	);
 
 	add_settings_field(
-		'gutenberg-full-page-client-side-navigation',
-		__( 'iAPI: full page client side navigation', 'gutenberg' ),
-		'gutenberg_display_experiment_field',
-		'gutenberg-experiments',
-		'gutenberg_experiments_section',
-		array(
-			'label' => __( 'Enables full-page client-side navigation with the Interactivity API, updating HTML while preserving application state.', 'gutenberg' ),
-			'id'    => 'gutenberg-full-page-client-side-navigation',
-		)
-	);
-
-	add_settings_field(
 		'gutenberg-editor-write-mode',
 		__( 'Simplified site editing', 'gutenberg' ),
 		'gutenberg_display_experiment_field',